OSS-65: split internal packages up into multiple methods
